Refactor: move to modules and extract type folder and file
