+++ apps/api-server/src/app.module.ts
@@ -1,12 +1,12 @@
-import { AuthModule } from './auth/auth.module';
+import { AuthModule } from './modules/auth/auth.module';
 import { Module } from '@nestjs/common';
 
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ConfigModule } from './config/config.module';
 import { DatabaseModule } from './database/database.module';
-import { UsersModule } from './users/users.module';
-import { DrinksCategoryModule } from './drinks-category/drinks-category.module';
+import { UsersModule } from './modules/users/users.module';
+import { DrinksCategoryModule } from './modules/drinks-category/drinks-category.module';
 
 @Module({
 	imports: [ConfigModule, DatabaseModule, UsersModule, AuthModule, DrinksCategoryModule],
