+++ apps/api-server/src/entities/drinks-category.entity.ts
@@ -2,8 +2,7 @@ import { ApiProperty } from '@nestjs/swagger';
 import { IsNotEmpty, IsString } from 'class-validator';
 import { CommonEntity } from './common.entity';
 import { Column, Entity } from 'typeorm';
-import { Category } from '../drinks-category/drinks-category.types';
-
+import { Category } from '../types/drinks-category.types';
 @Entity()
 export class DrinksCategory extends CommonEntity {
 	@ApiProperty({
