+++ apps/api-server/src/modules/auth/auth.service.ts
@@ -4,7 +4,7 @@ import { JwtConfig, OauthConfig } from '@src/config/config.constant';
 import { ConfigService } from '@nestjs/config';
 import { JwtService } from '@nestjs/jwt';
 import { Repository } from 'typeorm';
-import { User } from '../entities/users.entity';
+import { User } from '../../entities/users.entity';
 import { UserKakaoDto } from './dtos/users.kakao.dto';
 import { hash } from 'bcrypt';
 import { InjectRepository } from '@nestjs/typeorm';
