+++ apps/api-server/src/modules/users/users.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from '../entities/users.entity';
+import { User } from '../../entities/users.entity';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
